fix: incorrectly inferred index type (#80)
In some cases where inputs can be possibly null or undefined, the
compiler gets confused computing the distributive union for the
conditional type and ends up generating an index signature for the
variable types. This then proceeds to interfere with typechecking for
the variables.

This fix works around the issue by forcing the compiler to "resolve" the
type completely which typically makes it evaluate to never and disappear
entirely - however, for the real null and undefined cases, we still
effectively generate `{}` as it should be with other atomic types,
preserving correctness.

Fixes #79 

side note: This does make me think that magically inferring variable
types is too much for the compiler. A simpler design where

```typescript
query({myVariable: $.myType, otherVariable: $.Array($.anotherType)}, (q, $) => [
  q.someQuery({someInput: $.myVariable, otherInput: $.otherVariable})
])
```

might not be as flashy and convenient, but may be a better option that
will significantly simplify the whole typing story, with `$.typename`
syntax ensuring that autocomplete is still relatively convenient for
selecting input types. But this would be a backward-incompatible change.