fix: variable inference under ts 5.7 (#86)
Under compiler version 5.7.2 (or more specifically since PR
https://github.com/microsoft/TypeScript/pull/52095) TypeScript will more
eagerly instantiate generic types when intersections are involved. This
causes all variables in the generic input type in TGB to be inferred as
having type `unknown`
https://github.com/typed-graphql-builder/typed-graphql-builder/issues/85

To prevent this from happening, we insert a dummy conditional type that
involves the generic type. This causes the compiler to defer
instantiating the `ExactArgNames` restriction and instead use other
types to do the generic inference, restoring correct variable type
inference.

I feel that at this point we're too reliant on implementation details of
the typescript compiler, and while this patch fixes the immediate
problem, a response from the TypeScript team would be much appreciated
to make better decisions on the direction going forward.